Soroban: move storage array push into codegen (arrays.rs)
- Add arrays.rs with soroban_vec_push_back + soroban_storage_push, using the
  storage encode form (soroban_storage_encode_arg) so the path is
  forward-compatible with composite elements.
- Add storage_array_push_i32_test in the new storage_array_batch.rs.
