+++ src/codegen/targets/soroban/mod.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod arrays;
 pub(crate) mod dispatch;
 pub(crate) mod encoding;
 pub(crate) mod events;
@@ -182,7 +183,7 @@ impl TargetCodegen for SorobanTarget {
         if elem_is_ref {
             storage_slots_array_push(loc, args, cfg, contract_no, func, ns, vartab, opt, self)
         } else {
-            soroban_storage_push(loc, args, cfg, contract_no, func, ns, vartab, opt, self)
+            arrays::soroban_storage_push(loc, args, cfg, contract_no, func, ns, vartab, opt, self)
         }
     }
 
@@ -1297,7 +1298,7 @@ fn push_codegen_error(ns: &mut Namespace, err: CodegenError) {
     }
 }
 
-fn soroban_vec_handle_ty(vec_ty: &Type) -> Type {
+pub(crate) fn soroban_vec_handle_ty(vec_ty: &Type) -> Type {
     let inner_ty = if let Type::StorageRef(_, inner) = vec_ty {
         inner.as_ref().clone()
     } else {
@@ -1337,40 +1338,6 @@ pub(crate) fn soroban_vec_new(
     empty_vec_var
 }
 
-fn soroban_vec_push_back(
-    loc: &pt::Loc,
-    vec_obj: Expression,
-    vec_ty: &Type,
-    value: Expression,
-    cfg: &mut ControlFlowGraph,
-    ns: &Namespace,
-    vartab: &mut Vartable,
-) -> Expression {
-    let value_encoded = soroban_encode_arg(value, cfg, vartab, ns);
-    let handle_ty = soroban_vec_handle_ty(vec_ty);
-
-    let new_vec_no = vartab.temp_name("soroban_vec_push", &handle_ty);
-
-    let new_vec_var = Expression::Variable {
-        loc: *loc,
-        ty: handle_ty.clone(),
-        var_no: new_vec_no,
-    };
-
-    let instr = Instr::Call {
-        res: vec![new_vec_no],
-        return_tys: vec![handle_ty],
-        call: InternalCallTy::HostFunction {
-            name: HostFunctions::VecPushBack.name().to_string(),
-        },
-        args: vec![vec_obj, value_encoded],
-    };
-
-    cfg.add(vartab, instr);
-
-    new_vec_var
-}
-
 fn soroban_vec_pop_back(
     loc: &pt::Loc,
     vec_obj: Expression,
@@ -1399,47 +1366,6 @@ fn soroban_vec_pop_back(
     cfg.add(vartab, instr);
 
     new_vec_var
-}
-
-pub(crate) fn soroban_storage_push(
-    loc: &pt::Loc,
-    args: &[ast::Expression],
-    cfg: &mut ControlFlowGraph,
-    contract_no: usize,
-    func: Option<&Function>,
-    ns: &Namespace,
-    vartab: &mut Vartable,
-    opt: &Options,
-    target: &dyn TargetCodegen,
-) -> Expression {
-    // Storage wrapper: evaluate storage key/value and load vec object from storage.
-    let var_expr = expression(&args[0], cfg, contract_no, func, ns, vartab, opt, target);
-    let value = expression(&args[1], cfg, contract_no, func, ns, vartab, opt, target);
-    let vec_ty = args[0].ty();
-
-    let old_vec_obj = load_storage(
-        loc,
-        &vec_ty,
-        var_expr.clone(),
-        cfg,
-        vartab,
-        None,
-        ns,
-        target,
-    );
-    let new_vec_var = soroban_vec_push_back(loc, old_vec_obj, &vec_ty, value, cfg, ns, vartab);
-
-    // Storage wrapper: store updated vec object.
-    let store_instr = Instr::SetStorage {
-        ty: vec_ty,
-        value: new_vec_var.clone(),
-        storage: var_expr.clone(),
-        storage_type: None,
-    };
-
-    cfg.add(vartab, store_instr);
-
-    var_expr
 }
 
 pub(crate) fn soroban_storage_pop(
@@ -2007,7 +1933,7 @@ pub(crate) fn soroban_default_handle(
     }
 }
 
-fn load_raw_handle(
+pub(crate) fn load_raw_handle(
     loc: &pt::Loc,
     storage: Expression,
     cfg: &mut ControlFlowGraph,
