+++ tests/soroban_testcases/mod.rs
@@ -27,6 +27,7 @@ mod print;
 mod sha256;
 mod storage;
 mod storage_array;
+mod storage_array_batch;
 mod strings;
 mod struct_abi_member_count;
 mod struct_abi_single_field;
